+++ test_strutil.c
@@ -3,7 +3,7 @@
 #include <stdlib.h>
 
 static void test_strtolh(CuTest *tc) {
-    CuAssertIntEquals(tc, 1000, strtolh("1K", 10));
+    CuAssertIntEquals(tc, 1000, (int)strtolh("1K", 10));
 }
 
 void add_suite_strutil(CuSuite *suite)
